[issue/12-add_news] refactor, rename packages
